requirements.txt
requests
beautifulsoup4
lxml
mysql-connector-python
python-dotenv
selenium
webdriver-manager
